git-svn-id: file:///home/danlipsa/dan-svn-repo/src@2794 d596a3c1-de11-4375-84b4-10211c464620

+++ Simulation.cpp
@@ -558,7 +558,7 @@ void Simulation::ParseTopologicalChanges (
     if (! in)
 	ThrowException ("Cannot open \"" + fileName + "\"");
     in.exceptions (ios::badbit);
-    size_t timeStep;
+    float timeStep;
     float x, y, z;
     size_t type;
     while (! in.eof ())
